Refactor listing creation to reusable ListingForm
Extracted the multi-step listing creation form into a new reusable ListingForm component to support both creation and editing of listings. Added an edit listing page that loads existing data and reuses the form. Updated ad.service and types to support new API calls and richer data structures.

## Melhorias a Implementar no Formulário de Anúncio

- **Upload de fotos**
  - Implementar envio real
  - Adicionar preview no passo de revisão

- **Feedback visual**
  - Adicionar loaders/spinners durante carregamentos

- **Mensagens**
  - Substituir `alert` por toasts para mensagens de sucesso/erro

- **Payload da API**
  - Incluir campos faltantes: `furnished`, `pets`, `smoking`, `universidade`

- **Tipagem**
  - Melhorar a tipagem do estado do formulário

- **Redirecionamento**
  - Redirecionar após sucesso na criação ou edição

- **Tratamento de erros**
  - Exibir mensagens mais detalhadas para falhas da API

- **Confirmação de saída**
  - Adicionar alerta ao tentar sair com dados não salvos

- **Máscaras de input**
  - Implementar máscaras para campos como **CEP** e **valor monetário**

- **Validação de fotos**
  - Validar formato e número de fotos enviadas

- **Edição de fotos**
  - Adicionar opção para remover fotos existentes
